Fix deletion of last package
